Extract functions of MasterComment class into own files
The MasterComment file contained a couple of functions.
This commit extracts the functions into own files and loads them
into the global scope.

- FetchDataRow => inc/Functions/FetchDataRow.php
- FetchPostRow => inc/Functions/FetchPostRow.php
- MasterCommentEditAllowed => inc/Functions/MasterCommentEditAllowed.php
