+++ inc/Functions/MasterCommentEditAllowed.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @return bool
+ */
+function MasterCommentEditAllowed() {
+    global $line, $auth;
+
+    if ($line['creatorid'] == $auth['userid'] || $auth['type'] >= 2) {
+        return true;
+    }
+
+    return false;
+}
