+++ inc/classes/class_mastercomment.php
@@ -1,61 +1,5 @@
 <?php
 
-/**
- * @param string $username
- * @return string
- */
-function FetchDataRow($username) {
-    global $func, $dsp, $line;
-
-    $html_image= '<img src="%s" alt="%s" border="0">';
-    $avatar = ($func->chk_img_path($line['avatar_path'])) ? sprintf($html_image, $line['avatar_path'], t('Avatar')) : '';
-
-    if ($line['userid']) {
-        $ret = $dsp->FetchUserIcon($line['userid'], $username);
-
-    } else {
-        $ret = '<i>'. t('Gast') .'</i>';
-    }
-
-    $ret .= HTML_NEWLINE;
-    $ret .= $func->unixstamp2date($line['date'], "datetime") . HTML_NEWLINE;
-    if ($avatar) {
-        $ret .= $avatar . HTML_NEWLINE;
-    }
-
-    return $ret;
-}
-
-/**
- * @param string $text
- * @return string
- */
-function FetchPostRow($text) {
-    global $func, $line;
-
-    $ret = '<span id="post' . $line['commentid'] . '">' . $func->text2html($text) . '</span>';
-    if ($line['signature']) {
-        $ret .= '<hr size="1" width="100%" color="cccccc">';
-        $ret .= $func->text2html($line['signature']);
-    }
-
-    return $ret;
-}
-
-/**
- * @return bool
- */
-function MasterCommentEditAllowed() {
-    global $line, $auth;
-
-    if ($line['creatorid'] == $auth['userid'] || $auth['type'] >= 2) {
-        return true;
-    }
-
-    return false;
-}
-
-
 class Mastercomment {
 
     /**
